requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
django-sslserver==0.22
djangorestframework==3.15.2
numpy==2.1.2
opencv-python==4.10.0.84
pillow==11.0.0
pymemcache==4.0.0
python-memcached==1.62
pyzbar==0.1.9
sqlparse==0.5.1
typing_extensions==4.12.2
